Add issue templates and configuration for GitHub
This patch adds some simple issue templates and contact links for
Upstream issue trackers such as Element's issue tracker as well as
flatpak's issue tracker.

This should provide some useful navigation for users and it should be
possible to provide an issue tracker URL in the flatpak itself that
allows users to find their way around.